requirement.txt
streamlit
pandas
numpy
scikit-learn
httpx
python-dotenv
langchain-openai
